refactoring reservation page
